Refactor hub http handlers
