+++ cmd/hub/handlers/chartrepo/handlers_test.go
@@ -1,0 +1,326 @@
+package chartrepo
+
+import (
+	"context"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"strings"
+	"testing"
+
+	"github.com/artifacthub/hub/internal/api"
+	"github.com/artifacthub/hub/internal/hub"
+	"github.com/artifacthub/hub/internal/tests"
+	"github.com/rs/zerolog"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+)
+
+func TestMain(m *testing.M) {
+	zerolog.SetGlobalLevel(zerolog.Disabled)
+	os.Exit(m.Run())
+}
+
+func TestAdd(t *testing.T) {
+	dbQuery := "select add_chart_repository($1::uuid, $2::text, $3::jsonb)"
+
+	t.Run("invalid chart repository provided", func(t *testing.T) {
+		testCases := []struct {
+			description string
+			repoJSON    string
+		}{
+			{
+				"no chart repository provided",
+				"",
+			},
+			{
+				"invalid json",
+				"-",
+			},
+			{
+				"missing name",
+				`{"url": "https://repo1.url"}`,
+			},
+			{
+				"missing url",
+				`{"name": "repo1"}`,
+			},
+			{
+				"invalid name",
+				`{"name": "1repo"}`,
+			},
+			{
+				"invalid name",
+				`{"name": "repo_underscore"}`,
+			},
+			{
+				"invalid name",
+				`{"name": "REPO_UPPERCASE"}`,
+			},
+		}
+		for _, tc := range testCases {
+			tc := tc
+			t.Run(tc.description, func(t *testing.T) {
+				hw := newHandlersWrapper()
+
+				w := httptest.NewRecorder()
+				r, _ := http.NewRequest("POST", "/", strings.NewReader(tc.repoJSON))
+				r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+				hw.h.Add(w, r)
+				resp := w.Result()
+				defer resp.Body.Close()
+
+				assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
+			})
+		}
+	})
+
+	t.Run("valid chart repository provided", func(t *testing.T) {
+		repoJSON := `
+		{
+			"name": "repo1",
+			"display_name": "Repository 1",
+			"url": "https://repo1.url"
+		}
+		`
+		testCases := []struct {
+			description        string
+			dbResponse         interface{}
+			expectedStatusCode int
+		}{
+			{
+				"success",
+				nil,
+				http.StatusOK,
+			},
+			{
+				"database error",
+				tests.ErrFakeDatabaseFailure,
+				http.StatusInternalServerError,
+			},
+		}
+		for _, tc := range testCases {
+			tc := tc
+			t.Run(tc.description, func(t *testing.T) {
+				hw := newHandlersWrapper()
+				hw.db.On("Exec", dbQuery, mock.Anything, mock.Anything, mock.Anything).
+					Return(tc.dbResponse)
+
+				w := httptest.NewRecorder()
+				r, _ := http.NewRequest("POST", "/", strings.NewReader(repoJSON))
+				r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+				hw.h.Add(w, r)
+				resp := w.Result()
+				defer resp.Body.Close()
+
+				assert.Equal(t, tc.expectedStatusCode, resp.StatusCode)
+				hw.db.AssertExpectations(t)
+			})
+		}
+	})
+}
+
+func TestDelete(t *testing.T) {
+	dbQuery := "select delete_chart_repository($1::uuid, $2::text)"
+
+	t.Run("valid request", func(t *testing.T) {
+		hw := newHandlersWrapper()
+		hw.db.On("Exec", dbQuery, mock.Anything, mock.Anything).Return(nil)
+
+		w := httptest.NewRecorder()
+		r, _ := http.NewRequest("DELETE", "/", nil)
+		r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+		hw.h.Delete(w, r)
+		resp := w.Result()
+		defer resp.Body.Close()
+
+		assert.Equal(t, http.StatusOK, resp.StatusCode)
+		hw.db.AssertExpectations(t)
+	})
+
+	t.Run("database error", func(t *testing.T) {
+		hw := newHandlersWrapper()
+		hw.db.On("Exec", dbQuery, mock.Anything, mock.Anything).Return(tests.ErrFakeDatabaseFailure)
+
+		w := httptest.NewRecorder()
+		r, _ := http.NewRequest("DELETE", "/", nil)
+		r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+		hw.h.Delete(w, r)
+		resp := w.Result()
+		defer resp.Body.Close()
+
+		assert.Equal(t, http.StatusInternalServerError, resp.StatusCode)
+		hw.db.AssertExpectations(t)
+	})
+}
+
+func TestGetOwnedByOrg(t *testing.T) {
+	dbQuery := "select get_org_chart_repositories($1::uuid, $2::text)"
+
+	t.Run("valid request", func(t *testing.T) {
+		hw := newHandlersWrapper()
+		hw.db.On("QueryRow", dbQuery, mock.Anything, mock.Anything).Return([]byte("dataJSON"), nil)
+
+		w := httptest.NewRecorder()
+		r, _ := http.NewRequest("GET", "/", nil)
+		r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+		hw.h.GetOwnedByOrg(w, r)
+		resp := w.Result()
+		defer resp.Body.Close()
+		h := resp.Header
+		data, _ := ioutil.ReadAll(resp.Body)
+
+		assert.Equal(t, http.StatusOK, resp.StatusCode)
+		assert.Equal(t, "application/json", h.Get("Content-Type"))
+		assert.Equal(t, tests.BuildCacheControlHeader(0), h.Get("Cache-Control"))
+		assert.Equal(t, []byte("dataJSON"), data)
+		hw.db.AssertExpectations(t)
+	})
+
+	t.Run("database error", func(t *testing.T) {
+		hw := newHandlersWrapper()
+		hw.db.On("QueryRow", dbQuery, mock.Anything, mock.Anything).Return(nil, tests.ErrFakeDatabaseFailure)
+
+		w := httptest.NewRecorder()
+		r, _ := http.NewRequest("GET", "/", nil)
+		r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+		hw.h.GetOwnedByOrg(w, r)
+		resp := w.Result()
+		defer resp.Body.Close()
+
+		assert.Equal(t, http.StatusInternalServerError, resp.StatusCode)
+		hw.db.AssertExpectations(t)
+	})
+}
+
+func TestGetOwnedByUser(t *testing.T) {
+	dbQuery := "select get_user_chart_repositories($1::uuid)"
+
+	t.Run("valid request", func(t *testing.T) {
+		hw := newHandlersWrapper()
+		hw.db.On("QueryRow", dbQuery, mock.Anything).Return([]byte("dataJSON"), nil)
+
+		w := httptest.NewRecorder()
+		r, _ := http.NewRequest("GET", "/", nil)
+		r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+		hw.h.GetOwnedByUser(w, r)
+		resp := w.Result()
+		defer resp.Body.Close()
+		h := resp.Header
+		data, _ := ioutil.ReadAll(resp.Body)
+
+		assert.Equal(t, http.StatusOK, resp.StatusCode)
+		assert.Equal(t, "application/json", h.Get("Content-Type"))
+		assert.Equal(t, tests.BuildCacheControlHeader(0), h.Get("Cache-Control"))
+		assert.Equal(t, []byte("dataJSON"), data)
+		hw.db.AssertExpectations(t)
+	})
+
+	t.Run("database error", func(t *testing.T) {
+		hw := newHandlersWrapper()
+		hw.db.On("QueryRow", dbQuery, mock.Anything).Return(nil, tests.ErrFakeDatabaseFailure)
+
+		w := httptest.NewRecorder()
+		r, _ := http.NewRequest("GET", "/", nil)
+		r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+		hw.h.GetOwnedByUser(w, r)
+		resp := w.Result()
+		defer resp.Body.Close()
+
+		assert.Equal(t, http.StatusInternalServerError, resp.StatusCode)
+		hw.db.AssertExpectations(t)
+	})
+}
+
+func TestUpdate(t *testing.T) {
+	dbQuery := "select update_chart_repository($1::uuid, $2::jsonb)"
+
+	t.Run("invalid chart repository provided", func(t *testing.T) {
+		testCases := []struct {
+			description string
+			repoJSON    string
+		}{
+			{
+				"no chart repository provided",
+				"",
+			},
+			{
+				"invalid json",
+				"-",
+			},
+		}
+		for _, tc := range testCases {
+			tc := tc
+			t.Run(tc.description, func(t *testing.T) {
+				hw := newHandlersWrapper()
+
+				w := httptest.NewRecorder()
+				r, _ := http.NewRequest("PUT", "/", strings.NewReader(tc.repoJSON))
+				r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+				hw.h.Update(w, r)
+				resp := w.Result()
+				defer resp.Body.Close()
+
+				assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
+			})
+		}
+	})
+
+	t.Run("valid chart repository provided", func(t *testing.T) {
+		repoJSON := `
+		{
+			"display_name": "Repository 1 updated",
+			"url": "https://repo1.url/updated"
+		}
+		`
+		testCases := []struct {
+			description        string
+			dbResponse         interface{}
+			expectedStatusCode int
+		}{
+			{
+				"success",
+				nil,
+				http.StatusOK,
+			},
+			{
+				"database error",
+				tests.ErrFakeDatabaseFailure,
+				http.StatusInternalServerError,
+			},
+		}
+		for _, tc := range testCases {
+			tc := tc
+			t.Run(tc.description, func(t *testing.T) {
+				hw := newHandlersWrapper()
+				hw.db.On("Exec", dbQuery, mock.Anything, mock.Anything).Return(tc.dbResponse)
+
+				w := httptest.NewRecorder()
+				r, _ := http.NewRequest("PUT", "/", strings.NewReader(repoJSON))
+				r = r.WithContext(context.WithValue(r.Context(), hub.UserIDKey, "userID"))
+				hw.h.Update(w, r)
+				resp := w.Result()
+				defer resp.Body.Close()
+
+				assert.Equal(t, tc.expectedStatusCode, resp.StatusCode)
+				hw.db.AssertExpectations(t)
+			})
+		}
+	})
+}
+
+type handlersWrapper struct {
+	db *tests.DBMock
+	h  *Handlers
+}
+
+func newHandlersWrapper() *handlersWrapper {
+	db := &tests.DBMock{}
+	hubAPI := api.New(db, nil)
+
+	return &handlersWrapper{
+		db: db,
+		h:  NewHandlers(hubAPI),
+	}
+}
