+++ cmd/hub/handlers/static/testdata/image.png
