+++ internal/tests/email.go
@@ -1,9 +1,14 @@
 package tests
 
 import (
+	"errors"
+
 	"github.com/artifacthub/hub/internal/email"
 	"github.com/stretchr/testify/mock"
 )
+
+// ErrFakeEmailSenderFailure represents a fake email sender failure.
+var ErrFakeEmailSenderFailure = errors.New("fake email sender failure")
 
 // EmailSenderMock is a mock implementation of the EmailSender interface.
 type EmailSenderMock struct {
